pyproject.toml
[tool.poetry]
name = "housing_regression"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
pandas = "^1.3.2"
llvmlite = "^0.37.0"
shap = "^0.39.0"
matplotlib = "^3.4.3"
pandasgui = "^0.2.13"
scikit-learn = "^0.24.2"
yellowbrick = "^1.3.post1"
spyder-kernels = ">=2.0.1,<2.1.0"
"mega.py" = "^1.0.8"
GitPython = "^3.1.20"
streamlit = "^0.88.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
